requirements.txt
Django==3.0.2
djangorestframework==3.11.0
pycparser==2.19
Pygments==2.6.1
django-cors-headers